[agent_builder] - refactor: move AgentAccessPublicationDialog to settings folder
 - Adjust imports in AgentBuilderLeftPanel to reflect the new file location
 - Organize the AgentAccessPublicationDialog into the settings module, improving codebase structure
